Attempt Juggernaut reconnect
This uses the same ws_key so should attempt to reestablish if the TCP
connection breaks. If I deliberately provoke the server into closing
the connection, by failing to send an ack, then it doesn't let me
re-establish with the same key. But we do now close the connection
and Pidgin will reopen a new one.