Use registered_auto_complates in send_im to unknown user
There also seems to be a contact_auto_completes but that isn't necessarily
what we want since if it's in our contacts we'd have found it already.

We could *perhaps* go as far as sending an invite to a new user if even
this fails, but I don't know how to do that without also adding them to
the buddy list.