Escape incoming IM messages
Someone sent me a patch for this. I saw it this morning when there had not
been sufficient caffeine. And now I cannot find it anywhere. Maybe there
still hasn't been sufficient caffeine. Easier just to fix it again myself.

Apologies for the lack of attribution to... whoever sent me the one that
prompted me to fix it.