Address ChimeRoom object lifetime
Rooms remain in the connection's hash table(s) until their ->dispose()
is called. However, the object may continue to exist after we've been
removed from the room (because there's an open chat using it). So in
that case we set a ->is_dead flag, and it remains in the hash table but
*without* owning a ref, so that the last g_object_unref() will let it die.

This way, when we're removed from and subsequently added back to the same
room, any extant ChimeRoom object keeps working (and we don't add a *new*
ChimeRoom object for the same room, after failing to find the old one
in the hash table).