Update room information on Room jugg message
This doesn't really change much for us at the moment, since we're not
paying attention to any of the fields which change. But theoretically
I believe we can be a lot more stateful than we are, and keep the room
and conversation data in local storage. We should be able to rely on
getting "replayed" juggernaut messages on reconnect, to update things
as necessary. In the case of room mentions it looks like we don't get
a replay of the RoomMessage that we missed, but we *do* get a Room
notification carrying an updated LastMentioned field.

So if we do transition to being stateful (as I suspect we should, in
time), this is a step towards keeping that working.