Fix ChimeRoom unhashing on ->dispose()
In rooms_cb() when a room is no longer present we had this helpful comment:
	/* It'll remove itself from the hash table in its ->dispose() */

A fine plan, foiled only by the fact that there isn't a chime_room_dispose()
method. And if there was, it had no pointer to the ChimeConnection with
which it could have found the hash table and removed itself.

So add it. It makes me even less happy about the object refcounting we have
here, and the hacks around hashing and is_dead.

We really don't want the ChimeRoom to hold a ref to the ChimeConnection,
since for non-dead rooms the ChimeConnection holds a ref to the ChimeRoom.

But we do need chime_room_dispose() to be able to unhash it. So we keep
a pointer to the ChimeConnection in the ChimeRoom *only* while the latter
is hashed, knowing that before the ChimeConnection it destroyed it'll
unhash all the rooms as it destroys the hash table.