Reorder teardown
This makes it more obvious when object cleanup is/isn't working. By
the time we clean up contacts, all the ephemeral ones should be gone
and we should have only the "buddies". And by the time we clean up
the juggernaut, everything should be gone and there should only be
the device/profile/presence channels to unsubscribe.