Start to make sense of conversations.c vs. chime-conversation.c
Use ChimeConversation objects, handle actual messages using a hacked
variant of the original code. Next up, a sane ChimeMessages object
for collecting messages for ChimeRooms and ChimeConversations, and
fixing up some of the remaining incest between Purple bits and Chime
bits. This is kind of ugly but it basically works and can be cleaned
up piecemeal from here... and it's been sitting around half-complete
for a week or so, so it's about time it got pushed.