More chime_read_last_msg() cleanup
We were calling it twice in a row from on_chime_new_conversation() now, which
is silly. And pass the ChimeRoom/ChimeConversation into it, instead of a boolean
to indicate the type and the ID as a string. Which will make it closer to
chime_update_last_msg() when we eventually get round to making it a signal
on the same objects. Which won't be tonight...