Emit message signal for sent messages
I caught it earlier when the juggernaut disconnected just as I was sending
a message. Perhaps unsurprisingly, when you reconnect they don't bother
to replay your *own* messages to you, so my own message didn't appear in
the chat.

It occurs to me as I type this commit message that we might not see messages
which were sent by our *other* clients, during a period of disconnection.
Will need to check that. We should definitely be monitoring for LastSent
and triggering a message refetch if we notice that we're out of date.