Fix stack overflow when padding option array in fetch_room_memberships.
While-loop condition was "<= 4" meaning we actually ended up trying to
pad the array with 5 array values rather than 4 which lead to a stack
overflow and client crash. Initializing the array with NULL makes the
code agnostic to how many values it needs to pad preventing this from
happening again in the future.