Fix fetching of room messages
We can't rely on having a current LastSent for Rooms, so always fetch
from our last known message, just in case.

It would be nice if Chime always sent us a 'PreviousMessageId' with
each message, which would allow us to ditch this when we're opening
the room due to a mention. And would help in a few other cases too.