Update room names
We only get told of the name change on the room channel itself, which we don't
subscribe to unless we have it open — and even then it only tells us that it's
been updated; not what the change was. So we have to refetch it. But it's
better than nothing.

We should probably trigger a fetch_rooms() when the roomlist is brought up
in the UI, to make sure they're up to date.