Fix crash when multiple conversations exist with the same peer email address
There can be more than one profile for a given email address. You can have
conversations with both profiles. I don't know how this happens, but I've
seen the aftermath.

This was causing the first 'im' structure to be destroyed when the second
was discovered, which caused a use-after-free because I haven't yet
converted the chime_im/chime_room/chime_msgs stuff to proper GOBjects
and refcounted them, and the fetch_messages call was still pending.

For a quick "fix", keep them around by having a hash table on profile_id
too.