Explicitly close ChimeConnection in chime_purple_close()
Instead of allowing it to be implicitly closed by the final unref, close
it explicitly. This allows us to fix up the refcounting elsewhere, so
the objects associated with the connection can just hold a ref to it.

It was kind of OK for rooms/conversations but it's getting silly with
meetings adn calls.