Fix support for more than 100 contacts
Oh joy, there's a completely separate way to handle fetching of contacts
in stages. The 'NextToken' is in a special header instead of in the JSON
response, but ultimately it's very similar. Make fetch_contacts() work
like fetch_conversations() and fetch_rooms() do.

This fixes the disappearance of contacts, when you have more than 100.
Which was partly masked by the fact that we weren't deleting the buddies
from Pidgin.