Start cleaning up audio support, handle mute
There are locking issues here. The appsink_new_sample() function can get
invoked from another thread, which means it can race with the timer.

I think if the g_source_remove() in do_send_rt_packet() runs while a
timed_send_rt_packet() is waiting for the rt_lock, it can deadlock.

The simple fix is probably to shift all the actual packet processing into
a separate thread with a simple packet-sending loop which does both audio
and muted frames.

Or maybe we don't want a thread wakeup to add to the outbound packet
latency, and we'll allow appsink_new_sample() to send directly for itself
with some simple locking scheme between it and the thread.

The thread can then handle DTLS and auth and everything else sequentially,
which lets me do DTLS "natively" with GnuTLS and saves me from lamenting
the absence of GDtlsClientConnection too much.