Let fs-app-transmitter launch a pipeline by its description
Instead of hard-coded audiotestsrc/filesink, let the pipeline descriptions
get passed in in the FsCandidate's 'host' and 'username' fields, much as
the FsShmTransmitter does for the filenames.

This might not last, as we can only reference named elements in a pipeline
description if they were created in that same pipeline, AFAICT. So my plan
of pre-creating the appsrc and then referencing it from the description
probably isn't going to work.

Which means we might end up with a dedicated FsChimeTransmitter after all,
but hey, at least I have a better understanding of how some of this stuff
fits together...