Prevent double-replacements in mentions
Outgoing mentions replace a name with a reference to the Chime user.
Sometimes you have two users with the same name in a chat room, or two
people where one name is a prefix of another one. This could end up with
something like `<@id1|<@id2|Last, First>>`, which is broken.

This change makes sure to only replace at word borders and not if a `|`
is in front. This isn't perfect, but better than before.