Cope with RTP as application/x-srtp by depayloading for ourselves
Farstream's FsRtpConference always installs the srtpenc/srtpdec elements
even when there is no encryption. In that case they just pass packets
through unmodified. Unfortunately the caps are negotiated wrong — it's
still marked as application/x-srtp. This means that rtpopusdepay won't
accept it, for example.

https://bugs.freedesktop.org/show_bug.cgi?id=105193
https://bugzilla.gnome.org/show_bug.cgi?id=793704

Work around this by actually *asking* for SRTP even though that's not
what we need. And doing the depayload for ourselves.

Now this should work with and upgraded FarStream 0.2.8 and everything
else on Ubuntu 16.04 out of the box.