Slight cleanup for call media handling
The call is connecting before do_join_chat() is ever called, which means we
may miss some of the early state transitions. Try to cope with that.

We probably want to drive the call connection directly from chat code in
the end, but this will do for now.

Also decouple the media setup and connection; the call window in Pidgin
should now say 'Calling...' and then 'Call in progress' once it's connected.

Hopefully this helps to deal with some of the issues with slow call setup.
I *was* able to reproduce the ever-slower call connection but now I can't.
Not sure if that's just coincidence...