More audio cleanup
Fix up lifetime handling on the appsrc/appsink and uninstall the callbacks
when we're done.

Also fix up the timer handling for empty outbound packets. When the system
was slow we were sending 'muted' RT packets in between the audio frames,
so set the timers *only* when we're actually muted. If we fail to send
audio at other times, let it just do a red mic like it should.