Initial RPM spec file
This makes the assumption that the package will be called pidgin-chime because
this is what it's called in `configure.ac`.

The main package includes the pidgin icon, rather than putting it in a separate
package, because it's not harmful to put unused icons on the filesystem in the
case where someone is using libpurple but not pidgin.

The evolution plugin is put into a separate package to minimize dependencies in
the main package.  Can be built --without evolution.