Refactor the login/sign-in code.
To conform with the backend/frontend split of this package, the sign in code has
been refactored by separtaing the screen scrapping parts from the front
end (user interaction) parts.

The split between sign in provider types is no more, so all the login dirty code
lives together in chime/chime-signin.c.

Also, a new utility (chime-get-token) has been incorporated to ease the debug of
the sign in process, and also as an easy way to obtain a session token for Chime
bot developers.