Do not automatically connect upon object construction.
chime_connection_new() automatically calls chime_connection_connect(), but if
the session token is not given, there hasn't been any chance to connect the
"authenticate" signal in order to properly transfer control through the code.

Instead of implicitly connecting upon object creation, let the caller do it
explicitly.  The purple plugin is also changed to conform to that.