Drop libecal-1.2 BR for specfile
It's not clear why we have to change the actual pkgconfig module name
instead of just the version. But any attempt to BR one or the other
would be fairly distro-specific anyway, based on %_fedora. So we might
as well just *not* BR it, and depend on the equally distro-specific
knowledge that it's in the evolution-data-server-devel package along
with all the other things that we *do* BR.

Signed-off-by: David Woodhouse <dwmw@amazon.co.uk>