Fix use-after-free crash in message gathering
When cleanup_msgs() is called, there may be a pending fetch. We were
seeing crashes when fetch_msgs_cb() ended up using the freed msgs
structure. Let cleanup_msgs() do the freeing instead of the caller
(since we already have this nasty hack that the msgs struct needs
to be at the start of the chat/im structure. And then defer the
final free() until fetch_msgs_cb() is called, if necessary.