Fetch ancient messages in 2-week chunks
Because the server doesn't play them in chronological order and we have to
sort them, this takes a lot of memory if we do years worth of messages in
one hit. Instead, fetch a heuristically reasonable sized amount at a time.

Fix msg_newer() which was returning the wrong answer, while splitting
out msg_newer_than().