Add Polari's code
Polari's UI fits quite well my idea for an initial ui for chime.
I took Polari 3.28 because it builds on Ubuntu 18 and I am too
lazy to setup my jhbuild to build gjs, modzjs etc. In the future
once we have everything more or less working we would like to
take the changes from Polari master and merge it here.