packaging: templatize debian/control
Perform substitution of Build-Depends and Depends per Debian/Ubuntu
release.  Build and binary dependencies for several releases are
provided; additional releases are simple to add.

In order to generate the .deb packaging metadata, a workflow such as
the following can be used (note that this assumes apt >= 2.0, which
introduced the 'satisfy' subcommand; on earlier releases, use a
combination of dpkg-checkbuilddeps and manual apt invocations or
pbuilder(8)):

$ make -f debian/rules.maint debian/changelog
$ sudo apt satisfy --no-install-recommends build-essential \
  "$(grep-dctrl -n -s Build-Depends . debian/control)"
$ dpkg-buildpackage -uc -us

This should generate proper Debian binary and source packages
targeting the current distribution provided the following conditions
are met:
1. The build is happening within the target distribution
2. The make, git, lsb-release, and dctrl-tools are installed

debian/control is generated with Build-Depends and binary Depends
substitute from the appropriate debian/deps/build-* and
debian/deps/bin-* files, respectively.

debian/changelog is generated based on git tag data with various
release and version fields substituded with the appropriate values for
the current distribution environment (e.g. Ubuntu focal or Debian
buster).

Signed-off-by: Noah Meyerhans <nmeyerha@amazon.com>