Update login to use external web browser
The login process changed again and this time we really do need a browser.

Use purple_notify_uri() to spawn it, and now we get to specify our own
callback_uri for it to redirect to with the token. Use chime:// because
we already have the scripting set up to capture that.

This means we can fix the problem of multiple accounts with the chime:
scheme handler, by passing the devtoken of the account in question. We
can also pass the UI name so that we know which UI to re-enable it in.

Signed-off-by: David Woodhouse <dwmw@amazon.co.uk>