Kill parse_time() and replace it with iso8601_to_ms()
GTimeVal has been deprecated; we need to use GDateTime to parse ISO8601
strings. All we really want is milliseconds so just create a helper to
do that, and convert the one user of parse_time() to it for now. Other
callers of g_time_val_from_iso8601() will use it next...

Signed-off-by: David Woodhouse <dwmw@amazon.co.uk>