Set GLIB_VERSION_MIN_REQUIRED and GLIB_VERSION_MAX_ALLOWED
The minimum is fairly ancieht because we build some gtk-2.0 and EDS bits,
and perhaps we could do that *only* for those components as the rest of
the code is clean. But this is OK for now.

Add a maximum too, to help avoid using stuff which is too new to be usable
on ancient distributions we might still want to support. Like g_memdup2().

Signed-off-by: David Woodhouse <dwmw@amazon.co.uk>